requirements.txt
Django==1.4.2
PIL==1.1.7
South==0.7.6
Whoosh==2.4.1
argparse==1.2.1
distribute==0.6.28
dj-database-url==0.2.1
django-allauth==0.8.2
django-dajax==0.9.2
django-dajaxice==0.2
django-haystack==1.2.7
django-tastypie==0.9.11
django-uni-form==0.9.0
django-user-streams==0.5.0
httplib2==0.7.7
mimeparse==0.1.3
oauth2==1.5.211
psycopg2==2.4.5
python-dateutil==1.5
python-openid==2.2.5
requests==1.0.3
wsgiref==0.1.2
xlrd==0.8.0
yolk==0.4.3
